Allow reuse of curl handle
This improves the performance of mruby-curl by reusing a single curl
handle for multiple requests.  This allows curl to maintain persistent
connections across calls which eliminates delays from starting up a new
connection (TCP handshake, TLS handshake, TCP slow-start).

There is an API difference here, previously Curl was a module and now it
is an instance.  The module functions are re-implemented as class
methods to allow old examples to continue to work.